use worker signals to clear out cache on startup and shutdown
