+++ koku/koku/database.py
@@ -20,6 +20,9 @@ import os
 from django.conf import settings
 from django.contrib.postgres.fields.jsonb import KeyTextTransform
 from django.contrib.postgres.fields.jsonb import KeyTransform
+from django.db import connections
+from django.db import DEFAULT_DB_ALIAS
+from django.db.migrations.executor import MigrationExecutor
 from django.db.models import DecimalField
 from django.db.models.aggregates import Func
 
@@ -66,6 +69,22 @@ def config():
 
     database_cert = ENVIRONMENT.get_value("DATABASE_SERVICE_CERT", default=None)
     return _cert_config(db_config, database_cert)
+
+
+def check_migrations():
+    """
+    Check the status of database migrations.
+    The koku API server is responsible for running all database migrations.  This method
+    will return the state of the database and whether or not all migrations have been completed.
+    Hat tip to the Stack Overflow contributor: https://stackoverflow.com/a/31847406
+    Returns:
+        Boolean - True if database is available and migrations have completed.  False otherwise.
+    """
+    connection = connections[DEFAULT_DB_ALIAS]
+    connection.prepare_database()
+    executor = MigrationExecutor(connection)
+    targets = executor.loader.graph.leaf_nodes()
+    return not executor.migration_plan(targets)
 
 
 class JSONBBuildObject(Func):
